+++ changelog.md
@@ -6,3 +6,6 @@
 ### 11:14
 - 创建LuckyNodeStudy项目，增加案例一：人生的意义[study readline]，文件：node-study-whatIsYourLife.js
 
+### 15:40
+- 增加node-study-net-webserver.js简易web服务器示例
+- 增加node-study-WebSocket基于socket.io的websocket演示demo